Taishawnk-Portfolio_Project-master/requirements.txt
asgiref==3.7.2
boto3==1.26.151
botocore==1.29.151
braintree==4.20.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
distlib==0.3.6
Django==4.2.2
django-cors-headers==4.0.0
django-csp==3.7
django-storages==1.13.2
djangorestframework==3.14.0
filelock==3.12.1
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
mysqlclient==2.1.1
packaging==23.1
Pillow==9.5.0
platformdirs==3.5.3
pluggy==1.0.0
psycopg2==2.9.6
pytest==7.3.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16

